Update outdated comparison with Telepresence
A lot has happened since that comparison was written. Telepresence now have full support for Docker. It runs in a Docker container and uses a docker network. It also uses a Docker volume driver to mount remote filesystems in a way that doesn't affect the host.

In addition to that, Telepresence injects the sidecars needed to perform intercepts using a mutating webhook.

Here's an attempt to update the comparison.